third_party/WebKit/LayoutTests/http/tests/inspector-enabled/console-exception-source-url-expected.txt

CONSOLE ERROR: line 1: Uncaught ReferenceError: FAIL is not defined
Tests that when exception happens before inspector is open source url is correctly shown in console.

foo2.js:1 Uncaught ReferenceError: FAIL is not defined
TEST COMPLETE
